+++ tests/testthat/test-assets.R
@@ -2,15 +2,18 @@ test_that("default options", {
   withr::local_options(list(
     gargle_oauth_cache = NULL,
     gargle_oob_default = NULL, httr_oob_default = NULL,
+    gargle_oauth_client_type = NULL,
     gargle_oauth_email = NULL,
     gargle_verbosity   = NULL,
     gargle_quiet       = NULL
   ))
   expect_equal(gargle_oauth_cache(), NA)
-  if (is_rstudio_server()) {
+  if (is_rstudio_server() || is_google_colab()) {
     expect_true(gargle_oob_default())
+    expect_equal(gargle_oauth_client_type(), "web")
   } else {
     expect_false(gargle_oob_default())
+    expect_equal(gargle_oauth_client_type(), "installed")
   }
   expect_null(gargle_oauth_email())
   expect_equal(gargle_verbosity(), "info")
@@ -30,6 +33,11 @@ test_that("gargle_oob_default() consults httr's option", {
     httr_oob_default = TRUE
   ))
   expect_true(gargle_oob_default())
+})
+
+test_that("gargle_oauth_client_type() consults the option", {
+  withr::local_options(list(gargle_oauth_client_type = "web"))
+  expect_equal(gargle_oauth_client_type(), "web")
 })
 
 test_that("gargle API key", {